fix: use the right path for the resource files

+++ pkg/addons/addon_rook.go
@@ -47,7 +47,7 @@ func (addon RookAddon) URL() string {
 func (addon RookAddon) Install(args ...string) {
 	node := *addon.masterNode
 
-	_, err := addon.communicator.RunCmd(node, "kubectl apply -f https://github.com/rook/rook/raw/master/cluster/examples/kubernetes/rook-operator.yaml")
+	_, err := addon.communicator.RunCmd(node, "kubectl apply -f https://raw.githubusercontent.com/rook/rook/master/cluster/examples/kubernetes/ceph/operator.yaml")
 	FatalOnError(err)
 	fmt.Println("waiting until rook is installed")
 	for {
@@ -57,15 +57,15 @@ func (addon RookAddon) Install(args ...string) {
 			break
 		}
 	}
-	_, err = addon.communicator.RunCmd(node, "kubectl apply -f https://github.com/rook/rook/raw/master/cluster/examples/kubernetes/rook-cluster.yaml")
+	_, err = addon.communicator.RunCmd(node, "kubectl apply -f https://raw.github.com/rook/rook/master/cluster/examples/kubernetes/ceph/cluster.yaml")
 	FatalOnError(err)
-	_, err = addon.communicator.RunCmd(node, "kubectl apply -f https://github.com/rook/rook/raw/master/cluster/examples/kubernetes/rook-storageclass.yaml")
+	_, err = addon.communicator.RunCmd(node, "kubectl apply -f https://raw.github.com/rook/rook/master/cluster/examples/kubernetes/ceph/storageclass.yaml")
 	FatalOnError(err)
-	_, err = addon.communicator.RunCmd(node, "kubectl apply -f https://github.com/rook/rook/raw/master/cluster/examples/kubernetes/rook-tools.yaml")
+	_, err = addon.communicator.RunCmd(node, "kubectl apply -f https://raw.github.com/rook/rook/master/cluster/examples/kubernetes/ceph/toolbox.yaml")
 	FatalOnError(err)
 	_, err = addon.communicator.RunCmd(node, "kubectl get storageclass | grep -v 'NAME' | awk '{print$1}' | xargs kubectl patch storageclass -p '{\"metadata\": {\"annotations\":{\"storageclass.kubernetes.io/is-default-class\":\"false\"}}}'")
 	FatalOnError(err)
-	_, err = addon.communicator.RunCmd(node, "kubectl patch storageclass rook-block -p '{\"metadata\": {\"annotations\":{\"storageclass.kubernetes.io/is-default-class\":\"true\"}}}'")
+	_, err = addon.communicator.RunCmd(node, "kubectl patch storageclass rook-ceph-block -p '{\"metadata\": {\"annotations\":{\"storageclass.kubernetes.io/is-default-class\":\"true\"}}}'")
 	FatalOnError(err)
 
 	fmt.Println("Rook installed")
@@ -75,16 +75,19 @@ func (addon RookAddon) Install(args ...string) {
 func (addon RookAddon) Uninstall() {
 	node := *addon.masterNode
 	addon.communicator.RunCmd(node, "kubectl delete -n rook pool replicapool")
-	addon.communicator.RunCmd(node, "kubectl delete storageclass rook-block")
+	addon.communicator.RunCmd(node, "kubectl delete storageclass rook-ceph-block")
 	addon.communicator.RunCmd(node, "kubectl delete crd clusters.rook.io pools.rook.io objectstores.rook.io filesystems.rook.io volumeattachments.rook.io  # ignore errors if on K8s 1.5 and 1.6")
 	addon.communicator.RunCmd(node, "kubectl delete -n rook-system daemonset rook-agent")
-	addon.communicator.RunCmd(node, "kubectl delete -f https://github.com/rook/rook/raw/master/cluster/examples/kubernetes/rook-operator.yaml")
+	addon.communicator.RunCmd(node, "kubectl delete -f https://raw.githubusercontent.com/rook/rook/master/cluster/examples/kubernetes/ceph/operator.yaml")
 	addon.communicator.RunCmd(node, "kubectl delete clusterroles rook-agent")
 	addon.communicator.RunCmd(node, "kubectl delete clusterrolebindings rook-agent")
 	time.Sleep(20 * time.Second)
 	addon.communicator.RunCmd(node, "kubectl delete namespace rook")
 
 	for _, node := range addon.nodes {
+		if node.IsEtcd || node.IsMaster {
+			continue
+		}
 		fmt.Printf("deleting rook on node %s\n", node.Name)
 		addon.communicator.RunCmd(node, "rm -rf /var/lib/rook")
 	}
